refactor Become tutor guide
