+++ apps/web/src/app/become-tutor/components/BecomeTutorGuide.tsx
@@ -2,205 +2,19 @@
 
 import { LoginButton } from '@mezon-tutors/app/components/auth/LoginButton';
 import {
-  ArrowRightLineIcon,
-  ClockOutlineIcon,
-  GrowthOutlineIcon,
-  WalletOutlineIcon,
-} from '@mezon-tutors/app/ui/icons';
-import {
   GUIDE_HIGHLIGHTS,
   GUIDE_STEPS,
   ROUTES,
-  type GuideHighlight,
-  type GuideHighlightIconKey,
-  type GuideStep,
 } from '@mezon-tutors/shared';
 import { useTranslations } from 'next-intl';
-import { useMedia, XStack, YStack, Paragraph, H1, H2, Text } from 'tamagui';
-
-type GuideStepCardProps = {
-  step: GuideStep;
-  isCompact: boolean;
-};
-
-type GuideHighlightCardProps = {
-  item: GuideHighlight;
-  isCompact: boolean;
-};
-
-const HIGHLIGHT_ICON_BY_KEY: Record<GuideHighlightIconKey, typeof WalletOutlineIcon> = {
-  setOwnRate: WalletOutlineIcon,
-  teachAnytime: ClockOutlineIcon,
-  growProfessionally: GrowthOutlineIcon,
-};
-
-function GuideStepCard({ step, isCompact }: GuideStepCardProps) {
-  const t = useTranslations('BecomeTutorGuide');
-
-  return (
-    <YStack
-      flex={1}
-      alignItems={isCompact ? 'flex-start' : 'center'}
-      paddingHorizontal={isCompact ? 8 : 12}
-      paddingVertical={isCompact ? 10 : 8}
-      gap="$2"
-    >
-      <XStack
-        width={52}
-        height={52}
-        borderRadius={14}
-        alignItems="center"
-        justifyContent="center"
-        backgroundColor="$myLessonsCategoryEnglishBackground"
-        borderWidth={1}
-        borderColor="$myLessonsCategoryEnglishDot"
-      >
-        <Text color="$myLessonsCategoryEnglishLabel" fontSize={27} fontWeight="800" lineHeight={27}>
-          {step.number}
-        </Text>
-      </XStack>
-      <H2 color="$myLessonsHeaderTitle" fontSize={isCompact ? 20 : 24} lineHeight={isCompact ? 24 : 30}>
-        {t(step.titleKey)}
-      </H2>
-      <Paragraph
-        color="$myLessonsPromoDescription"
-        fontSize={isCompact ? 12 : 11}
-        lineHeight={isCompact ? 17 : 16}
-        textAlign={isCompact ? 'left' : 'center'}
-        maxWidth={250}
-      >
-        {t(step.descriptionKey)}
-      </Paragraph>
-    </YStack>
-  );
-}
-
-function GuideHighlightCard({
-  item,
-  isCompact,
-}: GuideHighlightCardProps) {
-  const t = useTranslations('BecomeTutorGuide');
-  const Icon = HIGHLIGHT_ICON_BY_KEY[item.iconKey];
-
-  return (
-    <YStack
-      flex={1}
-      minHeight={isCompact ? 220 : 252}
-      padding={isCompact ? 16 : 18}
-      borderRadius={20}
-      borderWidth={1}
-      borderColor="$myLessonsCardBorder"
-      backgroundColor="$myLessonsCardBackground"
-      shadowColor="$myLessonsCardBorder"
-      shadowOpacity={0.12}
-      shadowRadius={14}
-      shadowOffset={{ width: 0, height: 10 }}
-      gap="$3"
-      position="relative"
-      overflow="hidden"
-      cursor="pointer"
-      group
-      hoverStyle={{
-        borderColor: '$myLessonsPrimaryButtonBorder',
-        backgroundColor: '$myLessonsPrimaryButton',
-        shadowColor: '$myLessonsPrimaryButton',
-        shadowOpacity: 0.34,
-        shadowRadius: 26,
-        y: -2,
-      }}
-    >
-      <YStack
-        position="absolute"
-        top={-34}
-        right={-30}
-        width={92}
-        height={92}
-        borderRadius={999}
-        backgroundColor="rgba(30, 97, 243, 0.08)"
-        $group-hover={{
-          backgroundColor: 'rgba(255,255,255,0.14)',
-        }}
-      />
-
-      <XStack
-        width={32}
-        height={32}
-        borderRadius={10}
-        alignItems="center"
-        justifyContent="center"
-        backgroundColor="rgba(32, 107, 255, 0.12)"
-        $group-hover={{
-          backgroundColor: 'rgba(255,255,255,0.16)',
-        }}
-      >
-        <Text
-          color="#1D66F2"
-          lineHeight={0}
-          $group-hover={{
-            color: '#F6FAFF',
-          }}
-        >
-          <Icon color="currentColor" />
-        </Text>
-      </XStack>
-
-      <H2
-        color="$myLessonsHeaderTitle"
-        fontSize={isCompact ? 30 : 38}
-        lineHeight={isCompact ? 35 : 44}
-        fontWeight="800"
-        $group-hover={{
-          color: '$myLessonsPrimaryButtonText',
-        }}
-      >
-        {t(item.titleKey)}
-      </H2>
-      <Paragraph
-        color="$myLessonsPromoDescription"
-        fontSize={12}
-        lineHeight={18}
-        opacity={1}
-        $group-hover={{
-          color: '$myLessonsPrimaryButtonText',
-          opacity: 0.92,
-        }}
-      >
-        {t(item.descriptionKey)}
-      </Paragraph>
-
-      <YStack flex={1} />
-
-      <XStack alignItems="center" justifyContent="space-between">
-        <Text
-          color="$myLessonsPrimaryButton"
-          fontSize={9}
-          letterSpacing={1.1}
-          fontWeight="700"
-          textTransform="uppercase"
-          $group-hover={{
-            color: '$myLessonsPrimaryButtonText',
-          }}
-        >
-          {t(item.tagKey)}
-        </Text>
-        <Text
-          color="$myLessonsPrimaryButton"
-          lineHeight={0}
-          $group-hover={{
-            color: '$myLessonsPrimaryButtonText',
-          }}
-        >
-          <ArrowRightLineIcon color="currentColor" />
-        </Text>
-      </XStack>
-    </YStack>
-  );
-}
+import { useMedia, XStack, YStack, Paragraph, H1 } from 'tamagui';
+import { GuideHighlightCard } from './GuideHighlightCard';
+import { GuideStepCard } from './GuideStepCard';
 
 export function BecomeTutorGuide() {
-  const t = useTranslations('BecomeTutorGuide');
-  const media = useMedia();
-  const isCompact = media.sm || media.xs;
+  const t = useTranslations('BecomeTutorGuide')
+  const media = useMedia()
+  const isCompact = media.sm || media.xs
 
   return (
     <YStack
@@ -264,10 +78,11 @@ export function BecomeTutorGuide() {
 
             <XStack flexDirection={isCompact ? 'column' : 'row'} gap={isCompact ? '$2.5' : '$2'}>
               {GUIDE_STEPS.map((step) => (
-                <GuideStepCard 
-                key={step.id} 
-                step={step} 
-                isCompact={isCompact} />
+                <GuideStepCard
+                  key={step.id}
+                  step={step}
+                  isCompact={isCompact}
+                />
               ))}
             </XStack>
           </YStack>
@@ -298,14 +113,15 @@ export function BecomeTutorGuide() {
         <YStack padding={isCompact ? 16 : 22} backgroundColor="$tutorsPageBackground" gap="$4">
           <XStack flexDirection={isCompact ? 'column' : 'row'} gap="$2.5">
             {GUIDE_HIGHLIGHTS.map((item) => (
-              <GuideHighlightCard 
-              key={item.id} 
-              item={item} 
-              isCompact={isCompact} />
+              <GuideHighlightCard
+                key={item.id}
+                item={item}
+                isCompact={isCompact}
+              />
             ))}
           </XStack>
         </YStack>
       </YStack>
     </YStack>
-  );
+  )
 }
